+++ src/Bladeburner/ui/GeneralActionElem.tsx
@@ -1,7 +1,10 @@
 import React, { useState } from "react";
 import { ActionTypes } from "../data/ActionTypes";
 import { createProgressBarText } from "../../../utils/helpers/createProgressBarText";
-import { formatNumber } from "../../../utils/StringHelperFunctions";
+import {
+    formatNumber,
+    convertTimeMsToTimeElapsedString,
+} from "../../../utils/StringHelperFunctions";
 import { IBladeburner } from "../IBladeburner";
 import { IPlayer } from "../../PersonObjects/IPlayer";
 
@@ -15,6 +18,20 @@ export function GeneralActionElem(props: IProps): React.ReactElement {
     const setRerender = useState(false)[1];
     const isActive = props.action.name === props.bladeburner.action.name;
     const computedActionTimeCurrent = Math.min(props.bladeburner.actionTimeCurrent+props.bladeburner.actionTimeOverflow, props.bladeburner.actionTimeToComplete);
+    const actionTime = (function(): number{
+        switch(props.action.name) {
+        case "Training":
+        case "Field Analysis":
+            return 30;
+        case "Diplomacy":
+        case "Hyperbolic Regeneration Chamber":
+            return 60;
+        case "Recruitment":
+            return props.bladeburner.getRecruitmentTime(props.player);
+        }
+        return -1; // dead code
+    })();
+    const successChance = props.action.name === "Recruitment" ? Math.max(0, Math.min(props.bladeburner.getRecruitmentSuccessChance(props.player), 1)) : -1;
 
     function onStart(): void {
         props.bladeburner.action.type = ActionTypes[(props.action.name as string)];
@@ -42,6 +59,10 @@ export function GeneralActionElem(props: IProps): React.ReactElement {
         </>}
         <br />
         <br />
-        <pre style={{display: 'inline-block'}} dangerouslySetInnerHTML={{__html: props.action.desc}}></pre>
+        <pre style={{display: 'inline-block'}} dangerouslySetInnerHTML={{__html: props.action.desc}}></pre><br /><br />
+        <pre style={{display: 'inline-block'}}>
+            Time Required: {convertTimeMsToTimeElapsedString(actionTime*1000)}
+            {successChance !== -1 && <><br />Estimated success chance: {formatNumber(successChance*100, 1)}%</>}
+        </pre>
     </>);
 }